Implement `add_assign_slice`
This is the main algorithm behind bignum addition. It assumes that
`a.len()` >= `b.len()`. If this assumption does not hold, `a` will
have to be extended first. This is handled elsewhere, since
`add_assign_slice` does not own its arguments.

In addition to these changes, this PR changes the API of
`add_with_carry` to better fit how it is used in `add_assign_slice`.