Refactor `add_assign_with_carry` (#3)
The current implementation of this function uses `overflowing_add` and
several branches. We expect this function to lie on a hot path, so we
should avoid branching whenever possible.

The new implementation of this functions takes 32-bit values as input,
and computes their sum using 64-bit arithmetic. This allows us to read
the carry digit directly from the result, which should be much faster.